* Update CMakeLists.txt to match the latest versions of libkqueue and libpthread_workqueue * Fixed some compilation errors on linux * Some progress regarding windows compatibility
git-svn-id: http://opensource.mlba-team.de/svn/xdispatch/trunk@92 9e4c620c-c391-4b63-a7fb-4924991ea895

+++ core/src/debug.c
@@ -89,7 +89,7 @@ void _dispatch_logv(const char *msg, va_list params){
             struct timeval tv;
             gettimeofday(&tv, NULL);
             fprintf(logfile, "=== log file opened for %s[%u] at %ld.%06u ===\n",
-                    getprogname() ?  getprogname() : "", getpid(), tv.tv_sec, tv.tv_usec);
+                    getprogname() ?  getprogname() : "", getpid(), tv.tv_sec, (unsigned int)tv.tv_usec);
         }
     }
     vfprintf(logfile, newbuf, params);
